vivi-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.116.1"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
loguru = "^0.7.3"
openai = "^1.98.0"
pydantic-settings = "^2.10.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
sqlalchemy = "^2.0.42"
alembic = "^1.16.4"
python-dotenv = "^1.1.1"
requests = "^2.32.4"
beautifulsoup4 = "^4.13.4"
playwright = "^1.54.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
